+++ tests/testthat/test-z-execute-mutation-simple.R
@@ -71,3 +71,51 @@ test_that("small counter", {
 
 
 })
+
+# Altered from: https://github.com/schloerke/gqlr/issues/8
+test_that("mutation works with names that are not query names", {
+
+  votes <- 3
+
+  "
+  type Query {
+    votes: Int!
+  }
+
+  type Mutation {
+    castVote: Boolean!
+  }
+
+  schema {
+    query: Query
+    mutation: Mutation
+  }
+  " %>%
+    gqlr_schema(
+      Query = function(...) {
+        list(votes = function(...) {
+          votes
+        })
+      },
+      Mutation = function(...) {
+        list(castVote = function(...) {
+          votes <<- 42
+          # Return TRUE to indicate that the mutation was successful
+          TRUE
+        })
+      }
+    ) ->
+  votes_schema
+
+  expect_votes_request <- function(...) {
+    expect_request(..., schema = votes_schema)
+  }
+
+  "{votes}" %>%
+    expect_votes_request('{ "data": { "votes": 3 } }')
+  "mutation Cast { castVote }" %>%
+    expect_votes_request('{ "data": { "castVote": true } }')
+  "{votes}" %>%
+    expect_votes_request('{ "data": { "votes": 42 } }')
+
+})